Added enum mdoc_disp (similar to enum mdoc_list).  Display types are now only calculated once in mdoc_validate.c.
Noted that `Bd -file xxx' is not supported:  it now raises a fatal
warning.  This is noted in mdoc.7.

Empty `Bd' now defaults to LIST_ragged, which is not quite what groff
does, but close enough (gross just throws away the `Bd' and gets upset
when it encounters an `Ed').